pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-paradime-dbt"
authors = [
    {name = "Bhuvan Singla", email = "bhuvan@paradime.io"},
]
license = {text = "Apache License 2.0"}
description = "This is provider for Paradime to run and manage dbt™ jobs in production. The provider enables interaction with Paradime’s Bolt scheduler and management APIs."
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = ["version"]
requires-python = "~=3.8"
dependencies = ["apache-airflow>=2.4"]

[project.urls]
Homepage = "https://paradime.io"
Source = "https://github.com/paradime-io/airflow-provider-paradime-dbt"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "paradime_dbt_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "paradime_dbt_provider.__version__"}

[tool.setuptools.packages.find]
exclude = ["*tests.*", "*tests"]

[tool.black]
line-length = 200

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 200
combine_as_imports = true
src_paths = ["paradime_dbt_provider"]

[tool.ruff.isort]
combine-as-imports = true
